Move `Box` to `shared`
