+++ src/frontend/src/modules/MiniGrid/BoxGroup.vue
@@ -47,7 +47,7 @@
 
 <script>
 import { MiniGridService } from "@/services/MiniGridService"
-import Box from "../Box"
+import Box from "@/shared/Box.vue"
 import { currency } from "@/mixins/currency"
 
 export default {
