+++ src/frontend/src/shared/Box.vue
